Remove All Animations Other Than Left And Right
I basically removed the AnimationTree node altogether and just flip the sprite depending on the direction of the character. The sprite flipping will be useful for the combat animations implemented later in development.

(P.S: Forget the my placeholder animations. What we have right now should suffice until we get the proper sprites.)